Reorganize io into subpackage
