+++ cooler/io/__init__.py
@@ -1,0 +1,20 @@
+# # Exports
+# from ._binning import (ContactBinner, HDF5Aggregator, TabixAggregator,
+#                        PairixAggregator, CoolerAggregator, CoolerMerger,
+#                        SparseLoader, BedGraph2DLoader, ArrayLoader,
+#                        sanitize_pixels, validate_pixels, sanitize_records, aggregate_records)
+
+# from ._writer import (write_chroms, write_bins, prepare_pixels, write_pixels, 
+#                       write_indexes, write_info, index_bins, index_pixels, 
+#                       MAGIC, URL, PIXEL_FIELDS, PIXEL_DTYPES)
+
+from .ingestion import (
+    HDF5Aggregator, TabixAggregator, PairixAggregator, ArrayLoader,
+    sanitize_pixels, validate_pixels, sanitize_records, aggregate_records,
+    BadInputError
+)
+from .creation import create, create_from_unordered, append
+from .reduction import CoolerMerger, CoolerAggregator, merge, coarsen, zoomify
+#from .ingestion import from_hiclib, from_tabix, from_pairix, from_arraylike
+
+from .utils import ls, cp, mv, is_cooler, parse_cooler_uri, rename_scaffolds #tree
